pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-analyzer"
version = "0.1.0"
description = "Extract video frames and perform chained OpenAI vision analysis"
readme = "README.md"
authors = [ { name = "Your Name" } ]
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["video", "openai", "vision", "analysis"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "opencv-python",
  "pillow",
  "openai",
  "python-dotenv"
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]

[project.urls]
Homepage = "https://example.com"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = false
